+++ Assets/Scripts/Finish.cs
@@ -1,24 +1,15 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Events;
 
 public class Finish : MonoBehaviour
 {
-    private Game game;
-    private CircleCollider2D myCollider;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        game = FindObjectOfType<Game>();
-        myCollider = GetComponent<CircleCollider2D>();
-    }
+    [SerializeField] UnityEvent OnFinishHasBeenCollided;
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
         if (collision.gameObject.CompareTag("Player"))
-        {
-            game.win();
-        }
+            OnFinishHasBeenCollided.Invoke();
     }
 }
